feat: Enable custom CSS classes, footnote divider, localisation, source maps
Add rendering options:

- `prefixData` used in the main data attribute name as `data-${prefixData}footnotes`. The default value `data-footnotes` can be seen too general, when a widget is included on a foreign page and doesn't want to accidentally conflict with its other parts.
- `footnoteDivider`: inserts a horizontal rule above the footnotes as it's usually seen in books.
- `sectionClass`: used in the `class` attribute on the `section` element with the footnotes as `class="{sectionClass}"`. The default value `footnotes` can be seen too general, when a widget is included on a foreign page and doesn't want to accidentally conflict with its other parts.
- `headingClass`: used in the `class` attribute on the `h2` element with the footnotes title as `class="{headingClass}"`. The default value `sr-only` can be seen too general, when a widget is included on a foreign page and doesn't want to accidentally conflict with its other parts.
- `backRefLabel`: enables localisation of the ARIA label of the back-to-reference links. The value should use `{0}` as a placeholder for the reference label similarly to the English default: `Back to reference {0}`.

Distribute source maps. If `marked` is used in the browser, they help when investigating problems with the debugger.

Introduce `build:dev` script to run the build with the minification disabled. It's useful for debugging if source maps interpretation fails, or if the build output is integrated in an application by copying the concatenated JavaScript module as-is, or wrapped in the UMD wrapper, serving as a source in another application-specific build.

Exclude `dist/` from `eslint` checks. Minified files somehow "magically" survived, but the not minified files started getting complaints. No need to run static code checks against the build output.